pyproject.toml
[project]
name = "alertmanager-webhook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jonathan Ballet", email = "jon@multani.info" }
]
requires-python = ">=3.13"
dependencies = [
    "aiohttp==3.13.3",
    "click==8.2.1",
    "httpx==0.28.1",
    "prometheus-client==0.23.1",
    "pydantic==2.12.5",
    "pyyaml==6.0.2",
    "slack-sdk==3.36.0",
    "starlette==0.47.2",
    "structlog==25.4.0",
    "uvicorn==0.35.0",
]

[project.scripts]
alertmanager-webhook = "alertmanager_webhook.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alertmanager_webhook"]

[dependency-groups]
dev = [
    "mypy==1.17.1",
    "pytest==9.0.2",
    "pytest-cov==7.0.0",
    "ruff==0.12.9",
    "types-pyyaml==6.0.12.20250915",
]

[tool.pytest]
addopts = [
    "--cov=alertmanager_webhook",
    "--cov-branch",
    "--cov-report=html",
    "--cov-report=term",
    "--doctest-modules",
    "--tb=short",
    "--verbose",
    "--verbose",
]
testpaths = [
    "NAME",
    "tests",
]

[tool.mypy]
files = [
    "src",
    "tests",
]
strict = true

[tool.ruff.lint]
ignore = [
    # It's OK to have long line
    "E501",
]
extend-select = [
    # Sort imports
    "I",
]

[tool.ruff.lint.isort]
force-single-line = true
